requirements.txt
appnope==0.1.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
fake-useragent==0.1.11
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.6
jupyter-core==4.6.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.6
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
pyzmq==19.0.2
requests==2.24.0
six==1.15.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
